Add support for Shopify's Hydrogen/Oxygen (#1184)
* feat(core): replace deprecated url parse function

* feat(core): mark project as composite

* fixup! feat(core): replace deprecated url parse function

* feat: replace fetch stand in

Replaces custom method with node-fetch.

* style: formatting

* feat(react): add to-error function

* refactor(core): fold requestUrl function

Just use fetch.

* fix(react): change definition of safeDynamicRequire

Changed the definition of `safeDynamicRequire` check if `require` is a
function, rather than assuming it is if `Builder.isServer` is true. This
is required for CloudFlare Workers, where it is a server environment,
but `require` is not defined.

* fix(react): change the definition of `fetch`

Change fetch to use `globalThis`, which is the standardized global
object in modern browsers and runtimes (browser, NodeJS, WebWorkers,
etc.) to check if fetch is defined.

* fix(react): fix error handling

* style: format