fix(sdk): properly write to the root state (instead of repeated state)
When adding an action to a repeated item, the action code assignment
would write to the local state. This is incorrect as local state is
pseudo-state created for the purposes of the repeater. The correct
behavior is to write to the actual state.

This is achieved by having a Proxy that intercepts the write and correctly
writes to the root state.