📦 Publish SDKs (#2896)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-qwik@0.7.6

### Patch Changes

- 992a97e: Fix: Show & Hide If properties when combined with repeated
elements

## @builder.io/sdk-react@0.7.6

### Patch Changes

- 992a97e: Fix: Show & Hide If properties when combined with repeated
elements

## @builder.io/sdk-react-native@0.7.6

### Patch Changes

- 992a97e: Fix: Show & Hide If properties when combined with repeated
elements

## @builder.io/sdk-solid@0.7.6

### Patch Changes

- 992a97e: Fix: Show & Hide If properties when combined with repeated
elements

## @builder.io/sdk-svelte@0.7.6

### Patch Changes

- 992a97e: Fix: Show & Hide If properties when combined with repeated
elements

## @builder.io/sdk-react-nextjs@0.7.6



## @builder.io/sdk-vue@0.7.6

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>