📦 Publish SDKs (#2907)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react@0.8.1

### Patch Changes

- f5ea331: Fix: dynamic import of `isolated-vm` to rely on
`createRequire` instead of `eval("require")`

## @builder.io/sdk-react-native@0.8.1

### Patch Changes

- f5ea331: Fix: dynamic import of `isolated-vm` to rely on
`createRequire` instead of `eval("require")`

## @builder.io/sdk-svelte@0.8.1

### Patch Changes

- f5ea331: Fix: dynamic import of `isolated-vm` to rely on
`createRequire` instead of `eval("require")`

## @builder.io/sdk-react-nextjs@0.8.1



## @builder.io/sdk-qwik@0.8.1



## @builder.io/sdk-solid@0.8.1



## @builder.io/sdk-vue@0.8.1

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>