📦 Publish SDKs (#2915)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react-nextjs@0.10.0

### Minor Changes

- 39149d5: 🧨 Breaking: `fetchAllEntries`/`getAllContent` now returns the
array of contents directly, instead of an object with a `results`
property.

## @builder.io/sdk-qwik@0.10.0

### Minor Changes

- 39149d5: 🧨 Breaking: `fetchAllEntries`/`getAllContent` now returns the
array of contents directly, instead of an object with a `results`
property.

## @builder.io/sdk-react@0.10.0

### Minor Changes

- 39149d5: 🧨 Breaking: `fetchAllEntries`/`getAllContent` now returns the
array of contents directly, instead of an object with a `results`
property.

### Patch Changes

-   39149d5: compile SDK to ES2019 to support webpack 4 more easily.

## @builder.io/sdk-react-native@0.10.0

### Minor Changes

- 39149d5: 🧨 Breaking: `fetchAllEntries`/`getAllContent` now returns the
array of contents directly, instead of an object with a `results`
property.

## @builder.io/sdk-solid@0.10.0

### Minor Changes

- 39149d5: 🧨 Breaking: `fetchAllEntries`/`getAllContent` now returns the
array of contents directly, instead of an object with a `results`
property.

## @builder.io/sdk-svelte@0.10.0

### Minor Changes

- 39149d5: 🧨 Breaking: `fetchAllEntries`/`getAllContent` now returns the
array of contents directly, instead of an object with a `results`
property.

## @builder.io/sdk-vue@0.10.0

### Minor Changes

- 39149d5: 🧨 Breaking: `fetchAllEntries`/`getAllContent` now returns the
array of contents directly, instead of an object with a `results`
property.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>