📦 Publish SDKs (#2922)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-vue@0.11.0

### Minor Changes

-   70fa50d: - 🧨 Breaking: removed Vue 2 SDK

- 🧨 Breaking: removed `@builder.io/sdk-vue/vue3` import. To import the
SDK, you should now use:

    ```ts
    // BEFORE
    import { Content } from '@builder.io/sdk-vue/vue3';

    // AFTER
    import { Content } from '@builder.io/sdk-vue';
    ```

## @builder.io/sdk-react-nextjs@0.11.0



## @builder.io/sdk-qwik@0.11.0



## @builder.io/sdk-react@0.11.0



## @builder.io/sdk-react-native@0.11.0



## @builder.io/sdk-solid@0.11.0



## @builder.io/sdk-svelte@0.11.0

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>