📦 Publish SDKs (#2942)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react-nextjs@0.11.5

### Patch Changes

- 9a631fa: Fix: update "/edge" and "/node" subpath exports to only point
to their corresponding bundles. This guarantees that the browser bundle
is never imported by mistake.

## @builder.io/sdk-react@0.11.5

### Patch Changes

- 9a631fa: Fix: update "/edge" and "/node" subpath exports to only point
to their corresponding bundles. This guarantees that the browser bundle
is never imported by mistake.

## @builder.io/sdk-vue@0.11.5

### Patch Changes

- 9a631fa: Fix: update "/edge" and "/node" subpath exports to only point
to their corresponding bundles. This guarantees that the browser bundle
is never imported by mistake.

## @builder.io/sdk-qwik@0.11.5



## @builder.io/sdk-react-native@0.11.5



## @builder.io/sdk-solid@0.11.5



## @builder.io/sdk-svelte@0.11.5

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>