📦 Publish SDKs (#3031)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react-nextjs@0.14.4

### Patch Changes

- 9b873cd: Feature: allow passing `search` param (of type
`URLSearchParams | string | object`) to `isPreviewing` and `isEditing`
helpers. This allows users to rely on this function in SSR environments
to determine whether the current request is a preview or edit request.

## @builder.io/sdk-qwik@0.14.4

### Patch Changes

- 9b873cd: Feature: allow passing `search` param (of type
`URLSearchParams | string | object`) to `isPreviewing` and `isEditing`
helpers. This allows users to rely on this function in SSR environments
to determine whether the current request is a preview or edit request.

## @builder.io/sdk-react@1.0.6

### Patch Changes

-   9b873cd: Fix: add missing helper exports from `/edge` subpath
-   9b873cd: Fix: broken `/edge` CJS entry point
- 9b873cd: Feature: allow passing `search` param (of type
`URLSearchParams | string | object`) to `isPreviewing` and `isEditing`
helpers. This allows users to rely on this function in SSR environments
to determine whether the current request is a preview or edit request.
-   9b873cd: Remove `/server` subpath export (no longer needed)

## @builder.io/sdk-react-native@1.0.6

### Patch Changes

- 9b873cd: Feature: allow passing `search` param (of type
`URLSearchParams | string | object`) to `isPreviewing` and `isEditing`
helpers. This allows users to rely on this function in SSR environments
to determine whether the current request is a preview or edit request.

## @builder.io/sdk-solid@0.14.4

### Patch Changes

- 9b873cd: Feature: allow passing `search` param (of type
`URLSearchParams | string | object`) to `isPreviewing` and `isEditing`
helpers. This allows users to rely on this function in SSR environments
to determine whether the current request is a preview or edit request.

## @builder.io/sdk-svelte@1.0.6

### Patch Changes

- 9b873cd: Feature: allow passing `search` param (of type
`URLSearchParams | string | object`) to `isPreviewing` and `isEditing`
helpers. This allows users to rely on this function in SSR environments
to determine whether the current request is a preview or edit request.

## @builder.io/sdk-vue@1.0.6

### Patch Changes

- 9b873cd: Feature: allow passing `search` param (of type
`URLSearchParams | string | object`) to `isPreviewing` and `isEditing`
helpers. This allows users to rely on this function in SSR environments
to determine whether the current request is a preview or edit request.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>