📦 Publish SDKs (#3040)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react@1.0.9

### Patch Changes

- d9668eb: Fix: use `eval("require")` in favor of `node:module` when
importing `isolated-vm` to avoid issues when bundling in Nextjs

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>