📦 Publish SDKs (#3047)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react-nextjs@0.14.6

### Patch Changes

- b81e35a: fix: Image block `role='presentation'` set when altText prop
is not provided.

## @builder.io/sdk-qwik@0.14.6

### Patch Changes

- b81e35a: fix: Image block `role='presentation'` set when altText prop
is not provided.

## @builder.io/sdk-react@1.0.11

### Patch Changes

- b81e35a: fix: Image block `role='presentation'` set when altText prop
is not provided.

## @builder.io/sdk-react-native@1.0.11

### Patch Changes

- b81e35a: fix: Image block `role='presentation'` set when altText prop
is not provided.

## @builder.io/sdk-solid@0.14.6

### Patch Changes

- b81e35a: fix: Image block `role='presentation'` set when altText prop
is not provided.

## @builder.io/sdk-svelte@1.0.11

### Patch Changes

- b81e35a: fix: Image block `role='presentation'` set when altText prop
is not provided.

## @builder.io/sdk-vue@1.0.11

### Patch Changes

- b81e35a: fix: Image block `role='presentation'` set when altText prop
is not provided.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>