📦 Publish SDKs (#3249)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react-nextjs@0.14.19

### Patch Changes

- bd21dcf: Fix: improve NodeJS runtime performance by reusing the same
IsolatedVM Isolate instance for all data bindings. Add the ability to
provide arguments to configure the isolate in `initializeNodeRuntime`
via an `ivmIsolateOptions` parameter.
- c4c8cc1: Feature: added `builderLinkComponent` and `builderComponents`
props to RSC blocks.

## @builder.io/sdk-qwik@0.14.19

### Patch Changes

- bd21dcf: Fix: improve NodeJS runtime performance by reusing the same
IsolatedVM Isolate instance for all data bindings. Add the ability to
provide arguments to configure the isolate in `initializeNodeRuntime`
via an `ivmIsolateOptions` parameter.

## @builder.io/sdk-react@1.0.25

### Patch Changes

- bd21dcf: Fix: improve NodeJS runtime performance by reusing the same
IsolatedVM Isolate instance for all data bindings. Add the ability to
provide arguments to configure the isolate in `initializeNodeRuntime`
via an `ivmIsolateOptions` parameter.

## @builder.io/sdk-react-native@1.0.25

### Patch Changes

- bd21dcf: Fix: improve NodeJS runtime performance by reusing the same
IsolatedVM Isolate instance for all data bindings. Add the ability to
provide arguments to configure the isolate in `initializeNodeRuntime`
via an `ivmIsolateOptions` parameter.

## @builder.io/sdk-solid@1.0.25

### Patch Changes

- bd21dcf: Fix: improve NodeJS runtime performance by reusing the same
IsolatedVM Isolate instance for all data bindings. Add the ability to
provide arguments to configure the isolate in `initializeNodeRuntime`
via an `ivmIsolateOptions` parameter.

## @builder.io/sdk-svelte@1.0.25

### Patch Changes

- bd21dcf: Fix: improve NodeJS runtime performance by reusing the same
IsolatedVM Isolate instance for all data bindings. Add the ability to
provide arguments to configure the isolate in `initializeNodeRuntime`
via an `ivmIsolateOptions` parameter.

## @builder.io/sdk-vue@1.0.25

### Patch Changes

- bd21dcf: Fix: improve NodeJS runtime performance by reusing the same
IsolatedVM Isolate instance for all data bindings. Add the ability to
provide arguments to configure the isolate in `initializeNodeRuntime`
via an `ivmIsolateOptions` parameter.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>