📦 Publish SDKs (#3451)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-angular@0.2.2

### Patch Changes

- bdd3762: Fix: `onInit` hook to run initialisations (converted to
`ngOnInit`) on both server and client and move `onMount` and `onUpdate`
hooks to only run on browser (converted to `ngOnInit` and `ngOnChanges`
respectively with browser check)

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>