EDU-83: code snippets update in section integration (#3488)
## Description

As per the JIRA ticket [EDU-83](https://builder-io.atlassian.net/browse/EDU-83), I have implemented all the necessary changes for the Section model across frameworks as per Steve's loom. 

With these changes, a section model like the Announcements Bar will no longer require the getStaticPaths function. As a result of our current code structure, instead of removing getStaticPaths for the NextJs pages router, we return an empty array of paths to avoid the obvious error – getStaticPaths is required for dynamic SSG pages.. This decision was confirmed during a discussion with @steve8708 and was based on a suggestion from [stackoverflow]((https://stackoverflow.com/questions/65783199/error-getstaticpaths-is-required-for-dynamic-ssg-pages-and-is-missing-for-xxx)) to prevent the error.


[EDU-83]:
https://builder-io.atlassian.net/browse/EDU-83?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Authored by: Uttej Venkata Sastry <uttej@builder.io>