📦 Publish SDKs (#3506)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react-native@2.0.7

### Patch Changes

- 6a4c1f5: Feature: add optional `strictStyleMode` prop to `<Content>`
that will validate and ignore any style that does not work on React
Native. The validation logic is identical to the one in the Visual
Editor's "Strict React Native Styling" advanced setting.
-   6a4c1f5: Fix: gracefully handle css sanitization errors.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>