📦 Publish SDKs (#3512)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react-native@2.0.10

### Patch Changes

- 561e375: Fix: Drag and drop blocks into the correct place and add
blocks using the "+Add Block" button

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>