📦 Publish SDKs (#3535)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-angular@0.2.10

### Patch Changes

- cb68f66: Fix: update exported `content-variants` selector to `content`

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>