📦 Publish SDKs (#3540)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-angular@0.2.11

### Patch Changes

- 6003607: Fix: support destructuring out objects in Angular. Fixes
slot, nested symbols and element event binding in blocks other than
button

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>