📦 Publish SDKs (#3596)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react-nextjs@0.16.12

### Patch Changes

-   c58c2f9: misc: inline `lru-cache` dependency
-   c58c2f9: Fix: previewing content while inline editing.
    Chore: refactored build process to `preserveModules` in output.

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>