📦 Publish SDKs (#3612)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk-react@2.0.19

### Patch Changes

- 9836542: Fix: remove node-runtime logic from browser and edge bundles

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>