📦 Publish SDKs (#3626)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/react@5.0.5

### Patch Changes

-   0fc86b4: Fix: server-side-rendering of dynamic style bindings

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>