docs(nuxt): use proper key to useAsyncData (#3702)
## Description

After struggling to understand why navigation was not working properly
on prerendered routes I finally understood what was the culprit.

Nuxt uses the key of `useAsyncData` to cache that value on the server
and the client. Prerendered routes will have their async data stored
inside a payload file that is shared to the client when navigating.

In the current documentation it uses a static key for all pages content:
```ts
const { data: content } = await useAsyncData('builderData', () =>
  fetchOneEntry({
    model,
    apiKey,
    userAttributes: {
      urlPath: route.path,
    },
  })
))
```

When routes are prerender, navigation will fetch the payload of the page
and replace async data on the client based on the key. If the key
already exist, it is not replaced. So in this case, navigating between
pages will always result with the content of the first visited page.

To avoid this issue, each page should have its own key:

```ts
const { data: content } = await useAsyncData(`builderData-page-${route.path}`, () =>
  fetchOneEntry({
    model,
    apiKey,
    userAttributes: {
      urlPath: route.path,
    },
  })
))
```

This PR update the documentation accordingly