fix[gen2-sdks]: accordion visual editing and order (#3717)
## Description

for visual editing, we needed to pass the correct path. In gen1, if
`component.options` is not already prefixed we add them to the path:
https://github.com/sidmohanty11/builder/blob/fix/accordion-gen2/packages/react/src/components/builder-blocks.component.tsx#L48
so passing just `item.$index.detail` didn't work

for misordering of item details, we should only use `order` when
`props.grid === true`:
https://github.com/sidmohanty11/builder/blob/fix/accordion-gen2/packages/widgets/src/components/Accordion.tsx/#L75-L80

Jira
https://builder-io.atlassian.net/browse/ENG-7330

_Screenshot_
If relevant, add a screenshot or two of the changes you made.