📦 Publish SDKs (#3720)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @builder.io/sdk@4.0.2

### Patch Changes

-   dda2ba4: Fix: Restrict event listening to when `isEditing === true`

## @builder.io/react@6.0.2

### Patch Changes

- dda2ba4: Fix: Add `trustedHost` checks to all remaining event
listeners
-   dda2ba4: Fix: Restrict event listening to when `isEditing === true`
-   Updated dependencies [dda2ba4]
    -   @builder.io/sdk@4.0.2

## @builder.io/sdk-angular@0.2.26

### Patch Changes

- 5e88efa: Logs every API URL hit from the SDK whenever
`process.env.DEBUG` is set to `true` in the project

## @builder.io/sdk-react-nextjs@0.16.23

### Patch Changes

- 5e88efa: Logs every API URL hit from the SDK whenever
`process.env.DEBUG` is set to `true` in the project

## @builder.io/sdk-qwik@0.16.23

### Patch Changes

- 5e88efa: Logs every API URL hit from the SDK whenever
`process.env.DEBUG` is set to `true` in the project

## @builder.io/sdk-react@2.0.30

### Patch Changes

- 5e88efa: Logs every API URL hit from the SDK whenever
`process.env.DEBUG` is set to `true` in the project

## @builder.io/sdk-react-native@2.0.30

### Patch Changes

- 5e88efa: Logs every API URL hit from the SDK whenever
`process.env.DEBUG` is set to `true` in the project

## @builder.io/sdk-solid@2.0.30

### Patch Changes

- 5e88efa: Logs every API URL hit from the SDK whenever
`process.env.DEBUG` is set to `true` in the project

## @builder.io/sdk-svelte@2.0.30

### Patch Changes

- 5e88efa: Logs every API URL hit from the SDK whenever
`process.env.DEBUG` is set to `true` in the project

## @builder.io/sdk-vue@2.0.30

### Patch Changes

- 5e88efa: Logs every API URL hit from the SDK whenever
`process.env.DEBUG` is set to `true` in the project

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>