[phrasePlugin][Stale Private Key]:  (#3755)
## Description
The Private key in phrase plugin gets updated only in constructor i.e.
when the plugin loads. This happens on initial set-up or when we refresh
builder web-app. This stale private key causes authorization failure
when we switch b/w environments.
The solution is a small fix which gets the latest private key everytime
before making api-call.

**Loom showing before after**

https://www.loom.com/share/63c07935720749b7800aaf422b771ae4

---------

Co-authored-by: Sanyam Kamat <1625114+sanyamkamat@users.noreply.github.com>