fix[sdks]: `locale` prop to resolve localized values correctly (#3762)
## Description

A user can use two ways to resolve a localized value: either by passing
it through fetchOneEntry/get/getAll and resolving it on the API level,
or by providing a `locale` prop which should do it for them. Turns out
the `locale` prop was not resolving to the correct values.

We receive localized values in this format:
![Screenshot 2024-12-03 at 3 53
02 PM](https://github.com/user-attachments/assets/93485a4b-9f82-47ed-b0ee-d67797955423)

which errors out as text is now an object.

### Approach:
If a user has not provided locale on the API fetch level then we receive
options having these localized objects so if a user has provided a
`locale` prop we resolve them to the correct value else we warn the
users that they may encounter issues.

Jira
https://builder-io.atlassian.net/browse/ENG-7798

_Screenshot_
If relevant, add a screenshot or two of the changes you made.