Live Previewing Data Model – Add Support for Latest subscribeToEditor() Change in React Gen2 Snippets (#3907)
## Description

As per the ticket
[EDU-459](https://builder-io.atlassian.net/browse/EDU-459), this PR
updates the React Gen2 code snippet to align with the latest breaking
change to `subscribeToEditor()`. This ensures that the docs remain up to
date and work without any issues.

ENG changelog:
https://github.com/BuilderIO/builder/blob/main/packages/sdks/output/react/CHANGELOG.md#400

This PR adds the code for live previewing data models in the code
snippets. I just wanted to let you know that more to come.
@samijaber Please review it.


[EDU-459]:
https://builder-io.atlassian.net/browse/EDU-459?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Uttej Venkata Sastry <uttej@builder.io>