fix[angular-gen2]: A/B tests visual editing (#3958)
## Description

getters in angular run on every change detection, this PR extracts out
the default content getter and updates it to a function. This is a
temporary fix to get the A/B tests visual editing working correctly but
in the future we need a way to update the getters to be actual computed
values (only dependent upon state/prop updates)

Jira
https://builder-io.atlassian.net/browse/ENG-8217
https://builder-io.atlassian.net/browse/ENG-8498

Slack thread

https://builder-internal.slack.com/archives/C07P1H7496H/p1741591102140589

_Screenshot_
If relevant, add a screenshot or two of the changes you made.