React Gen 2- Product Details - Blueprints snippet  (#3946)
## Description

This PR aims to add Product Details blueprints code to GitHub for React
Gen 2 SDK as part of this [epic in the
JIRA-138](https://builder-io.atlassian.net/browse/EDU-138)

@samijaber Please review it.

I have added more code to the `useEffect()`, which is only for cleanup.
I know this isn't very significant for this scale of code, but I thought
it would be better to use some best practices so that customers would
know our code is resilient.