fix[sdks]: centering items in columns (#3979)
## Description

when we vertically center two individual elements we use `margin-top:
auto`, `margin-bottom: auto` to keep the elements in the center. Turns
out the builder-columns box doesn't respect height and shrinks to what’s
required and height gets applied to the root div (as we don’t use noWrap
true in case of columns). To fix this, we are adding `height: 100%` to
make sure that we respect the root div's height (if provided)

Jira
https://builder-io.atlassian.net/browse/ENG-8386

_Screenshot_
If relevant, add a screenshot or two of the changes you made.