fix: [ENG-8644] overriding omit in fetchOneEntry to be empty string (#3975)
### **Description:**  
This PR fixes an issue where `meta.componentsUsed` is omitted by default
when using `fetchOneEntry`. Previously, to include
`meta.componentsUsed`, you had to explicitly set `omit: ' '` (a space)
as a workaround.

### **Changes Made:**  
- Updated the logic to include `meta.componentsUsed` by default unless
explicitly omitted.
- Added `omit ?? 'meta.componentsUsed'` to ensure `componentsUsed` is
not omitted when `omit` is set to an empty string (`''`).

### **Why This Change Was Made:**  
To allow `meta.componentsUsed` to be included without requiring a
workaround and to ensure better handling of the `omit` parameter.


__Loom__

https://www.loom.com/share/938cd062796c4e2aaaf42b408039713e?sid=f087a66d-478d-42ce-ae41-b8e809535da6

---------

Co-authored-by: Sami Jaber <me@sami.website>
Co-authored-by: Sidharth Mohanty <sidmohanty11@gmail.com>