EDU-548: Custom components - Child blocks - Nextjs 14 (#3980)
## Description

This is the final leg of adding code to GitHub for the custom
components—child blocks as per the ticket
[EDU-548](https://builder-io.atlassian.net/browse/EDU-548) and its
parent [EDU-133](https://builder-io.atlassian.net/browse/EDU-133).

Whats changed in this PR compared to other PRs?
I had to define `builderComponents` and set it `true` in the the
component registration file like this:

> shouldReceiveBuilderProps: { builderBlock: true, builderComponents:
true },

Then, i passed `builderComonents` as a prop to the `<blocks/>` to
resolve Hydration error that i have reported and on the browser, i got
an error: `Text component is not registered`.




**Acceptance criteria:**

- [x] - Made sure there were no unnecessary conditional checks in the
code.
- [x] - Verified that no unused types and types are not required.
- [x] - Confirmed that the code runs locally and passes all test cases
without errors.
- [x] - Checked for any linting errors.
- [x] - Made sure no dead code that does not add value to users goes
into PR, ensuring that its removal won’t impact overall functionality.
- [x] - Adding `react-sdk-next-pages` code to the PR.

[EDU-548]:
https://builder-io.atlassian.net/browse/EDU-548?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[EDU-133]:
https://builder-io.atlassian.net/browse/EDU-133?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ