fix[ENG-8003]: Changing the entry on a symbol does not update the preview window (#3985)
## Description

This PR fixes an issue where symbol content was not updating correctly.
The fix compares the previous symbol entry with the incoming symbol
entry—if they are not equal, it triggers `fetchSymbolContent` to ensure
the latest content is loaded.

**Changes Made:**  
- Added a comparison check between the previous and incoming symbol
entries.
- Triggered `fetchSymbolContent` if the entries differ to update the
content.

**Why This Change Was Made:**  
To ensure symbol content updates correctly when the symbol entry
changes, improving accuracy and consistency in the preview.


_Loom_

https://www.loom.com/share/032123ec4a7f449db7690eb9ca022f3e?sid=3964bfa8-ac48-428a-a376-91568c447213

---------

Co-authored-by: Sidharth Mohanty <sidmohanty11@gmail.com>