fix[gen2][fetchOneEntry]: ENG-9110 make fetchOneEntry() handle options.locale just like builder.get() handles options.locale (#4114)
## Description

FetchEntries - includes extra userAttribute: {locale: value} along with
the locale payload

`builder.getAll()` hits the Content API without the addition of `locale`
to `userAttributes`. See here:
https://www.loom.com/share/7117e8725b69400f98e02d823a901b7a

The change in this [PR](https://github.com/BuilderIO/builder/pull/3719)
was added to ensure `locale` is added to `userAttributes` for
`builder.get()` and not `builder.getAll()`

However, to ensure the same outcome In Gen2 SDK the
[change](https://github.com/BuilderIO/builder/pull/3719/files#diff-b4fb78a40f76e70808858993da491c522b4f567feefeb6c01b0ab5defac58569)
of adding `locale` to `userAttributes` is added in
`packages/sdks/src/functions/get-content/generate-content-url.ts` which
is used in **BOTH** `fetchOneEntry()` and `fetchEntries()`.

**PR that caused the bug**
https://github.com/BuilderIO/builder/pull/3719

**Link to JIRA ticket**
https://builder-io.atlassian.net/browse/ENG-9110

**Loom**
https://www.loom.com/share/a5d6f20989c24811adc35545e92ea315