Fix typos and improve README.md (#4117)
Fixes several typos and improves readability

## Description

This PR addresses minor typos and enhances readability by updating the
README to better adhere to grammar and documentation standards. It also
adds a live-updating star count badge. The License badge now links to
the LICENSE file. The "Try it out!" header is changed to "Sign up" for
scanability.

_(this is my very first PR ever; let me know if I can improve it! Thanks
😊)_

---------

Co-authored-by: Korey Kassir <korey@builder.io>