fix[angular-gen2]: only use the edge runtime when the edge build is imported (#4152)
## Description

As we are skipping hydration due to limitations with how we generate
dynamic components we have to run eval twice - once on the server and
once on the browser. The edge build was working and edge-runtime was
being used but on the browser load it was fallbacking to use the browser
runtime which was leading to the unsafe-eval errors. This PR fixes the
issue by using edge-runtime both in the server and the browser - we will
only use the edge build when users import it using:
```ts
import { Content, type BuilderContent } from "@builder.io/sdk-angular/bundle/edge";
```

Jira
https://builder-io.atlassian.net/browse/ENG-10322

Screenshot
<img width="2555" height="690" alt="image"
src="https://github.com/user-attachments/assets/618bc9b2-e530-4547-b504-16c4be1ff369"
/>