fix[react-gen1]: handle `fetchPriority` vs `fetchpriority` in different react versions (#4155)
## Description

Handles fetch priority casing based on React versions

> In a previous PR, https://github.com/vercel/next.js/pull/47302,
detection for fetchPriority assumed that
https://github.com/facebook/react/pull/25927 would land in react@18.3.0
because that was the react@canary version at the time. However, it
didn't land in react@18.3.0 and it is expected to land in react@19.0.0
due to the breaking change.
This means that users upgrading to react@18.3.0 will see a warning.

Inspiration from:
https://github.com/facebook/react/issues/28946
https://github.com/vercel/next.js/pull/65235
https://github.com/ascorbic/unpic-img/pull/644

_Screenshot_
If relevant, add a screenshot or two of the changes you made.